feat: add empty-view package
Adds empty view package that contains four kind of empty view
components. In order to have a consisitent user experience across all
our applications, we need to use these common components when there is
no data to present.
Also adds useMediaQuery hook which is used in `MainEmptyView`. And it
will be used when we implement the responsive design.