fix: node+browser compatible ES module import paths
The packages that export ES modules do not have a `.js` extension on the
import paths, which means that the import path does not match an actual
file path (since the latter do have a `.js` extension). This means they
cannot be properly resolved by Node.js or the browser.

However, since we're using TypeScript to transpile, and we're using the
legacy "node" module resolution, the import paths are preserved and end
up not having a `.js` extension, while the transpiled files to have it.

There are two main ways to deal with this:

1. Use TS configuration `{"module":"esnext","moduleResolution":"bundler"}`
   and use a bundler that produces import paths matching an actual file.
   Note: using TypeScript in this case would leave the imports as-is
   and require users to use a bundler, so it would not constitute an
   ES module that just works with node/browser, which is the problem we
   actually set out to solve.
2. Use TS configuration `{"module":"nodenext","moduleResolution":"nodenext"}`
   and use `{"type":"module"}` in package.json. This will instruct tsc to
   produce output that matches the package configuration, so ES module in
   this case.

The advantage with (1) is that it requires no addition of a `.js`
extension to the imports in the .ts(x) source files. Unfortunately this
requires TS 5 (we currently use TS 4) _and_ the use of a bundler to
build (which we currently do not use). The advantage with (2) is that it
requires no changes to how we build, but it does required adding a `.js`
extension to all import paths in the source code.

Since addition of a bundler adds a layer of complexity in maintenance
for a repository that should be easily accessible, it seems option (2) is
preferable (and also can be done without having to do a major TS ugprade
first).