chore: replace eslint/dprint with biome
Removes eslint and adds biome, which is a much faster linter
that can handle TypeScript out of the box, and also does formatting.

Removes circular dependencies related to importing elements into the
route map, but also route map into elements. This caused a problem
with vite's HMR depending on the exact order of the imports.
The solution here is to only map routes to elements where needed
(in this case only App.tsx).